Write the whole buffer to gocode tmpfile
When creating the temporary file from the current buffer, only the lines
starting at `bufnr` to the end of the file are included, causing a
number of lines to be skipped depending on the current buffer number.

The missing lines would cause a go parsing error on gocode, preventing
it from being able to identify the context of the cursor and therefore
provide possible completion suggestions.

Fixes https://github.com/prabirshrestha/asyncomplete-gocode.vim/issues/4